+++ NEWS.md
@@ -1,6 +1,12 @@
 # inlmisc 0.4.3.9000
 
-- Change R version requirement from `>= 3.2.0` to `>= 3.4.0`.
+- Tidy leaflet-search plugin code.
+
+- In `CreateWebMap` function, change URL for TNM basemap WMS services from
+  <https://basemap.nationalmap.gov/arcgis/services/USGSTopo/MapServer/WmsServer> to
+  <https://basemap.nationalmap.gov/ArcGIS/rest/services/USGSTopo/MapServer/tile/{z}/{y}/{x}>.
+
+- Change required R version from `>= 3.2.0` to `>= 3.4.0`.
 
 # inlmisc 0.4.3